fixbug: args ptr will invalid if the msg timeout. see http://rb.corp.taobao.com/r/7444/
git-svn-id: http://code.taobao.org/svn/tfs@149 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/tfs-dev-1.3.2/src/new_client/tfs_file.h
@@ -100,25 +100,25 @@ namespace tfs
       int get_size_for_rw_ex(const int64_t check_size, const int64_t count, int64_t& cur_size, bool& not_end, const int64_t per_size);
 
     private:
-      int do_async_request(const InnerFilePhase file_phase, const int64_t wait_id, const int32_t index);
+      int do_async_request(const InnerFilePhase file_phase, const uint16_t wait_id, const int32_t index);
       int do_async_response(const InnerFilePhase file_phase, message::Message* packet, const int32_t index);
 
-      int async_req_create_file(const int64_t wait_id, const int32_t index);
+      int async_req_create_file(const uint16_t wait_id, const int32_t index);
       int async_rsp_create_file(message::Message* packet, const int32_t index);
 
-      int async_req_read_file(const int64_t wait_id, const int32_t index);
+      int async_req_read_file(const uint16_t wait_id, const int32_t index);
       int async_rsp_read_file(message::Message* packet, const int32_t index);
 
-      int async_req_write_data(const int64_t wait_id, const int32_t index);
+      int async_req_write_data(const uint16_t wait_id, const int32_t index);
       int async_rsp_write_data(message::Message* packet, const int32_t index);
 
-      int async_req_close_file(const int64_t wait_id, const int32_t index);
+      int async_req_close_file(const uint16_t wait_id, const int32_t index);
       int async_rsp_close_file(message::Message* packet, const int32_t index);
 
-      int async_req_stat_file(const int64_t wait_id, const int32_t index);
+      int async_req_stat_file(const uint16_t wait_id, const int32_t index);
       int async_rsp_stat_file(message::Message* packet, const int32_t index);
 
-      int async_req_unlink_file(const int64_t wait_id, const int32_t index);
+      int async_req_unlink_file(const uint16_t wait_id, const int32_t index);
       int async_rsp_unlink_file(message::Message* packet, const int32_t index);
 
     public:
